call stmt.ColumnConverter when implemented by parent statement
According to the stdlib driver documentation
(https://github.com/golang/go/blob/bc51e930274a5d5835ac8797978afc0864c9e30c/src/database/sql/driver/driver.go#L385)
value checkers should be called in the following order:

> [..] stopping at the first found match: Stmt.NamedValueChecker,
> Conn.NamedValueChecker, Stmt.ColumnConverter,
> DefaultParameterConverter.

sqlmw was not calling Stmt.ColumnConverter when it was implemented.

This commit changes the behavior to call Stmt.ColumnConverter, if it is
implemented and the NamedValueCheckers are not implemented.
This is done by returning ErrSkip in wrappedStmt.CheckNamedValue() if
neither the parent statement nor the conn implements CheckNamedValue.
The sql package will call wrappedStmt.ColumnConverter() if ErrSkip was
returned.

wrappedStmt.CheckNamedValue() can not check only if the stmt
implements CheckNamedValue and return ErrSkip.
It must also call CheckNamedValue() on the connection if it
was not implemented for the stmt. This is because the stdlib sql package
calls wrappedStmt.CheckNamedValue() if is implemented on the stmt OR on
the connection.

The commit also adds a testcase to verify that ColumnConverter is
called.