support intercepting Rows.Close() calls
Add a new RowsClose() method to the interceptor interface to which
rows.Close() calls are forwarded.

Intercepting rows.Close() can be useful in tracing middlewares.
It allows to create a parent span when the db Rows object is
created, create child spans for operations on the Rows and finish the
parent span when rows.Close() is called.