Revert "call stmt.ColumnConverter when implemented by parent statement"
My change caused that driver.DefaultParameterConverter is not called
anymore sometimes.
For example when the lib/pq driver is used, uint32 args passed to
stmt.Exec() are not converted by driver.DefaultParameterConverter to
int64 anymore.
This causes that the lib/pq driver operation fails because the argument
is passed unconverted, as unsupported type.

I could not reproduce the behavior in a testcase with the fakedb driver.
The driver.DefaultParameterConverter was called and the type was passed
converted to stmt.Exec(), as expected.

I don't understand yet why this happens with the lib/pq driver with my
change.

Revert it for now. That the deprecated ColumnConverter is not called by
sqlmw, is a lesser issue compared to the one that was introduced by my
change.