interceptor: remove ctx parameter from db methods that do not have it
The middleware was storing the context that was passed to methods
creating Rows, Tx and Stmts in the wrapped versions of those.
The ctx was then forwarded to operations like RowsNext, RowsClose,
StmtClose.
The related methods do not have a context parameter in the related db
methods in the stdlib.
That ctx was stored in the wrapped objects is probably a remaining of
the instrumentedsql package, that this package is forked from.
For the instrumentedsql package it made sense to store the ctx in the
wrapped object, because it contained tracing spans that could then be
accessed in RowsNext, RowsClose, etc.

For sqlmw it does not.
That the method signatures differs and context from different methods
were forwarded can be confusing. It was not clear for the user which
context he received in those interceptor methods.
The signatures of DB methods of the interceptor should ideally be as
close as possible to those of the stdlib db object.