interceptor: remove query arg from StmtExecContext and StmtQueryContext
The middleware was storing the query that was passed to PrepareContext
in the wrapped Stmt and then passing it to StmtExecContext and
StmtQueryContext.

Remove the query argument.
It is a remaining of the instrumentedsql pkg, were the query was stored
and passed to annotate tracing spans with it.

Removing it has the advantage that the interface is more close to the db
object interface of the stdlib. The string parameter was also
undocumented in the interface, it was only possible to figure out what
it is by reading the source code.

If the user needs to forward custom data he can wrap Rows, Stmt, Tx on
his own or better the middleware can provide in the future a mechanism
to forward arbitrary user defined data between those calls.